requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
email-validator==2.1.0
python-dotenv==1.0.0
psycopg[binary]
psycopg2-binary
Werkzeug==2.3.8
gunicorn==21.2.0
